Have tide set a more useful status.
This presents two concerns:

1. When it is first run, almost all statuses for PRs not in the tide
pool will be wrong. This means that we will reset them all at once,
which will eat at least several hundred tokens. Once we are hitting the
main k8s repo this will eat up to a few thousand. That's within our
budget, but we need to be careful and not make multiple tweaks within an
hour. We can also consider rate-limiting the status-setting separately
from the rest.

2. Looking up the labels on PRs costs an additional amount per PR.